#285 Update to user model and added new traits
