+++ app/Models/User.php
@@ -5,16 +5,13 @@ namespace App\Models;
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
 
 use App\Traits\HasTestPrefix;
+use App\Traits\UserHelpersTrait;
+use App\Traits\UserRelationshipsTrait;
+use App\Traits\UserScopesTrait;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Cache;
 use Laravel\Cashier\Billable;
 use Laravel\Fortify\TwoFactorAuthenticatable;
 use Laravel\Sanctum\HasApiTokens;
@@ -41,6 +38,9 @@ class User extends Authenticatable
      * @use HasApiTokens<\Laravel\Sanctum\HasApiTokens>
      * @use Billable<\Laravel\Cashier\Billable>
      * @use HasTestPrefix<\App\Traits\HasTestPrefix>
+     * @use UserRelationshipsTrait<\App\Traits\UserRelationshipsTrait>
+     * @use UserHelpersTrait<\App\Traits\UserHelpersTrait>
+     * @use UserScopesTrait<\App\Traits\UserScopesTrait>
      */
     use HasFactory,
         Notifiable,
@@ -48,7 +48,10 @@ class User extends Authenticatable
         SoftDeletes,
         HasApiTokens,
         Billable,
-        HasTestPrefix;
+        HasTestPrefix,
+        UserRelationshipsTrait,
+        UserHelpersTrait,
+        UserScopesTrait;
 
     /**
      * The attributes that are mass assignable.
@@ -107,221 +110,6 @@ class User extends Authenticatable
         'deleted_at' => 'datetime',
         'restored_at' => 'datetime',
     ];
-
-    /**
-     * Get the role assigned to the user.
-     *
-     * @return BelongsTo<Role,User>
-     */
-    public function role(): BelongsTo
-    {
-        return $this->belongsTo(Role::class);
-    }
-
-    /**
-     * Determine whether the user is a super administrator.
-     *
-     * @return bool
-     */
-    public function isSuperAdmin(): bool
-    {
-        return $this->role_id === Role::ROLE_SUPER_ADMIN;
-    }
-
-    /**
-     * Determine whether the user is an administrator.
-     *
-     * @return bool
-     */
-    public function isAdmin(): bool
-    {
-        return $this->role_id === Role::ROLE_ADMIN;
-    }
-
-    /**
-     * Determine whether the user is a standard user.
-     *
-     * @return bool
-     */
-    public function isUser(): bool
-    {
-        return $this->role_id === Role::ROLE_USER;
-    }
-
-    /**
-     * Get the permissions assigned to the user via their role.
-     *
-     * @return Collection<int,string>
-     */
-    public function permissions(): Collection
-    {
-        return $this->role
-            ? $this->role->permissions->pluck('name')->unique()
-            : collect();
-    }
-
-    /**
-     * Get all permissions for the user.
-     *
-     * Results are cached for 60 minutes to improve performance.
-     *
-     * @return array<int,string>
-     */
-    public function getAllPermissions(): array
-    {
-        return Cache::remember(
-            "user_permissions_{$this->id}",
-            60,
-            fn () => $this->permissions()->toArray()
-        );
-    }
-
-    /**
-     * Determine whether the user has a given permission.
-     *
-     * @param  string $permission The permission name.
-     *
-     * @return bool
-     */
-    public function hasPermission(string $permission): bool
-    {
-        return in_array($permission, $this->getAllPermissions());
-    }
-
-    /**
-     * Determine whether the user has a given role.
-     *
-     * Accepts either a role ID or role name.
-     *
-     * @param  int|string $role The role ID or name.
-     *
-     * @return bool
-     */
-    public function hasRole(int|string $role): bool
-    {
-        if (! $this->role) {
-            return false;
-        }
-
-        if (is_int($role)) {
-            return $this->role->id === $role;
-        }
-
-        return $this->role->name === $role;
-    }
-
-    /**
-     * Clear the cached permissions for the user.
-     *
-     * @return void
-     */
-    public function clearPermissionCache(): void
-    {
-        Cache::forget("user_permissions_{$this->id}");
-    }
-
-    /**
-     * Get the deals owned by the user.
-     *
-     * @return HasMany<Deal>
-     */
-    public function deals(): HasMany
-    {
-        return $this->hasMany(Deal::class, 'owner_id');
-    }
-
-    /**
-     * Get the tasks assigned to the user.
-     *
-     * @return HasMany<Task>
-     */
-    public function tasks(): HasMany
-    {
-        return $this->hasMany(Task::class, 'assigned_to');
-    }
-
-    /**
-     * Get the notes created by the user.
-     *
-     * @return HasMany<Note>
-     */
-    public function notes(): HasMany
-    {
-        return $this->hasMany(Note::class, 'user_id');
-    }
-
-    /**
-     * Get the learnings associated with the user.
-     *
-     * Includes pivot data such as completion status and timestamps.
-     *
-     * @return BelongsToMany<Learning>
-     */
-    public function learnings(): BelongsToMany
-    {
-        return $this->belongsToMany(Learning::class)
-            ->using(LearningUser::class)
-            ->withPivot([
-                'is_complete',
-                'user_id',
-                'completed_at',
-                'is_test',
-                'meta',
-                'created_by',
-                'updated_by',
-            ])
-            ->withTimestamps();
-    }
-
-    /**
-     * Get all attachments associated with the user.
-     *
-     * @return MorphMany<Attachment>
-     */
-    public function attachment(): MorphMany
-    {
-        return $this->morphMany(Attachment::class, 'attachable');
-    }
-
-    /**
-     * Get all activities associated with the user.
-     *
-     * @return MorphMany<Activity>
-     */
-    public function activity(): MorphMany
-    {
-        return $this->morphMany(Activity::class, 'subject');
-    }
-
-    /**
-     * Get all tasks where the user is the taskable entity.
-     *
-     * @return MorphMany<Task>
-     */
-    public function tasking(): MorphMany
-    {
-        return $this->morphMany(Task::class, 'taskable');
-    }
-
-    /**
-     * Get all notes associated with the user as a notable entity.
-     *
-     * @return MorphMany<Note>
-     */
-    public function note(): MorphMany
-    {
-        return $this->morphMany(Note::class, 'notable');
-    }
-
-    /**
-     * Get the job title associated with the user.
-     *
-     * @return BelongsTo<JobTitle,User>
-     */
-    public function jobTitle(): BelongsTo
-    {
-        return $this->belongsTo(JobTitle::class, 'job_title_id');
-    }
 
     /**
      * Get the formatted user name.
